requirements.txt
jikanpy-v4
pandas
numpy
tensorflow
keras
beautifulsoup4
scikit-learn
matplotlib
wordcloud
sendgrid
apache-airflow
pickle
